Reverts the network json two step into an ordered read/write, avoiding data duplication.
