+++ src/rust/lqosd/src/anonymous_usage/mod.rs
@@ -74,7 +74,7 @@ fn anonymous_usage_dump() -> anyhow::Result<()> {
 
     data.git_hash = env!("GIT_HASH").to_string();
     data.shaped_device_count = SHAPED_DEVICES.read().unwrap().devices.len();
-    data.net_json_len = NETWORK_JSON.load().get_nodes_when_ready().len();
+    data.net_json_len = NETWORK_JSON.read().unwrap().get_nodes_when_ready().len();
 
     data.high_watermark_bps = (
         HIGH_WATERMARK.get_down(),
