+++ src/rust/lqosd/src/node_manager/local_api/network_tree.rs
@@ -3,7 +3,7 @@ use lqos_config::NetworkJsonTransport;
 use crate::shaped_devices_tracker::NETWORK_JSON;
 
 pub async fn get_network_tree() -> Json<Vec<(usize, NetworkJsonTransport)>> {
-    let net_json = NETWORK_JSON.load();
+    let net_json = NETWORK_JSON.read().unwrap();
     let result: Vec<(usize, NetworkJsonTransport)> = net_json
         .get_nodes_when_ready()
         .iter()
